refactor(query): extract provider-option helpers into runner/provider_options.rs
