+++ src-rust/crates/query/src/runner/mod.rs
@@ -1,0 +1,4 @@
+// Runner submodule: cohesive helpers extracted from run_query_loop's
+// enclosing file (issue #232). Behavior-preserving moves.
+pub(crate) mod provider_options;
+pub(crate) use provider_options::*;
